added exception handler and  more controllers

+++ src/main/java/ru/softdepot/controllers/ProductsController.java
@@ -2,7 +2,6 @@ package ru.softdepot.controllers;
 
 import lombok.AllArgsConstructor;
 import org.springframework.http.HttpStatus;
-import org.springframework.http.HttpStatusCode;
 import org.springframework.http.ResponseEntity;
 import org.springframework.validation.BindException;
 import org.springframework.validation.BindingResult;
@@ -12,7 +11,6 @@ import ru.softdepot.Messages.Message;
 import ru.softdepot.core.dao.ProgramDAO;
 import ru.softdepot.core.models.Program;
 
-import java.util.List;
 import java.util.Map;
 
 @RestController
@@ -22,19 +20,27 @@ public class ProductsController {
     private final ProgramDAO programDAO;
 
     @GetMapping("/{id}")
-    public Program findProgram(@PathVariable("id") int id) throws Exception {
-        return this.programDAO.getById(id);
+    public ResponseEntity<?> findProgram(@PathVariable("id") int id) throws Exception {
+        if (!programDAO.exists(id))
+            return ResponseEntity
+                    .status(HttpStatus.NOT_FOUND)
+                    .body(Message.build(
+                            Message.Entity.product,
+                            Message.Identifier.id,
+                            id,
+                            Message.Status.notFound
+                    ));
+
+        var program = this.programDAO.getById(id);
+        return ResponseEntity.ok().body(program);
     }
 
     @PatchMapping("/edit")
     public ResponseEntity<?> updateProgram(@RequestBody Program program,
                                            BindingResult bindingResult) throws BindException {
         if (bindingResult.hasErrors()) {
-            if (bindingResult instanceof BindException exception) {
-                throw exception;
-            } else {
-                throw new BindException(bindingResult);
-            }
+            if (bindingResult instanceof BindException exception) throw exception;
+            else throw new BindException(bindingResult);
         } else {
             this.programDAO.update(program);
             return ResponseEntity.ok().build();
@@ -43,6 +49,16 @@ public class ProductsController {
 
     @DeleteMapping("/delete/{id}")
     public ResponseEntity<?> deleteProgram(@PathVariable("id") int id) {
+        if (!programDAO.exists(id))
+            return ResponseEntity
+                    .status(HttpStatus.NOT_FOUND)
+                    .body(Message.build(
+                            Message.Entity.product,
+                            Message.Identifier.id,
+                            id,
+                            Message.Status.notFound
+                    ));
+
         this.programDAO.delete(id);
         return ResponseEntity.ok().build();
     }
@@ -51,7 +67,7 @@ public class ProductsController {
     @PostMapping("/new")
     public ResponseEntity<?> addNewProgram(@RequestBody Program program,
                                            BindingResult bindingResult,
-                                           UriComponentsBuilder uriComponentsBuilder) throws BindException {
+                                           UriComponentsBuilder uriComponentsBuilder) throws Exception {
         if (bindingResult.hasErrors()) {
             if (bindingResult instanceof BindException exception) throw exception;
             else throw new BindException(bindingResult);
@@ -80,16 +96,8 @@ public class ProductsController {
                         ));
 
 
-
             int id = this.programDAO.add(program);
-            Program programResult;
-            try {
-                programResult = this.programDAO.getById(id);
-            } catch (Exception e) {
-                return ResponseEntity
-                        .badRequest()
-                        .body(e.getMessage() + "\n\n" + e.getStackTrace());
-            }
+            Program programResult = this.programDAO.getById(id);
             return ResponseEntity
                     .created(uriComponentsBuilder
                             .replacePath("/softdepot-api/products/{id}")
